feat(mac): regroup Settings by task and hide passive listening
Split crowded General and Audio panes into focused pages, trim Meetings copy, and keep passive/speaker enrollment out of the shipping UI until it is ready.
