+++ StetMac/Features/MacShell/AudioSetting/MacMicrophoneSettingsView.swift
@@ -1,0 +1,25 @@
+#if os(macOS)
+    import SwiftUI
+
+    struct MacMicrophoneSettingsView: View {
+        @StateObject private var viewModel = MacAudioSettingsViewModel()
+
+        var body: some View {
+            Form {
+                AudioInputDeviceSettingsSection(
+                    deviceManager: viewModel.deviceManager,
+                    microphoneTestViewModel: viewModel.microphoneTestViewModel
+                )
+            }
+            .formStyle(.grouped)
+            .padding(.leading, MacUI.SettingsViewMetrics.formHorizontalPadding)
+            .padding(.bottom, MacUI.SettingsViewMetrics.formBottomPadding)
+            .task {
+                viewModel.onAppear()
+            }
+            .onDisappear {
+                viewModel.onDisappear()
+            }
+        }
+    }
+#endif
