+++ StetMac/Features/MacShell/GeneralSetting/MacGeneralSettingsView.swift
@@ -10,7 +10,6 @@
         var body: some View {
             Form {
                 appBehaviorSection
-                dictationSection
                 updatesSection
                 #if DEBUG
                     debugSection
@@ -34,23 +33,6 @@
                 Text("Application")
             } footer: {
                 Text("Stet stays in the menu bar for quick access even when the dock icon is hidden.")
-            }
-        }
-
-        private var dictationSection: some View {
-            Section {
-                Toggle("Interaction sounds", isOn: $viewModel.interactionSoundsEnabled)
-                Toggle(
-                    "Notify when dictation completes",
-                    isOn: $viewModel.dictationCompletionNotificationsEnabled
-                )
-                Toggle("Mute background audio", isOn: $viewModel.pauseMediaDuringDictation)
-            } header: {
-                Text("Dictation")
-            } footer: {
-                Text(
-                    "Stet can play a sound and show a notification after your words are inserted."
-                )
             }
         }
