+++ StetMac/Features/MacShell/GeneralSetting/MacGeneralSettingsViewModel.swift
@@ -37,32 +37,6 @@
             }
         }
 
-        @Published var pauseMediaDuringDictation = false {
-            didSet {
-                guard hasLoadedPreferences else { return }
-                defaults.set(pauseMediaDuringDictation, forKey: MacPreferences.pauseMediaDuringDictation)
-            }
-        }
-        @Published var interactionSoundsEnabled = true {
-            didSet {
-                guard hasLoadedPreferences else { return }
-                defaults.set(interactionSoundsEnabled, forKey: MacPreferences.interactionSoundsEnabled)
-            }
-        }
-        @Published var dictationCompletionNotificationsEnabled = true {
-            didSet {
-                guard hasLoadedPreferences else { return }
-                defaults.set(
-                    dictationCompletionNotificationsEnabled,
-                    forKey: MacPreferences.dictationCompletionNotificationsEnabled
-                )
-                if dictationCompletionNotificationsEnabled {
-                    Task {
-                        await MacDictationCompletionNotificationService.shared.requestAuthorizationIfNeeded()
-                    }
-                }
-            }
-        }
         @Published var managedSettings = ManagedSettingsState() {
             didSet {
                 guard hasLoadedManagedSettings else { return }
@@ -86,7 +60,6 @@
         private weak var appModel: (any MacGeneralSettingsAppModeling)?
         private weak var appUpdateManager: AppUpdateManager?
 
-        private var hasLoadedPreferences = false
         private var hasLoadedManagedSettings = false
         private var hasLoadedDebugSettings = false
         private var suppressLaunchAtLoginChange = false
@@ -119,14 +92,6 @@
         }
 
         func load() {
-            hasLoadedPreferences = false
-            pauseMediaDuringDictation =
-                defaults.object(forKey: MacPreferences.pauseMediaDuringDictation) as? Bool ?? false
-            interactionSoundsEnabled = defaults.object(forKey: MacPreferences.interactionSoundsEnabled) as? Bool ?? true
-            dictationCompletionNotificationsEnabled =
-                defaults.object(forKey: MacPreferences.dictationCompletionNotificationsEnabled) as? Bool ?? true
-            hasLoadedPreferences = true
-
             hasLoadedManagedSettings = false
             managedSettings = currentState()
             hasLoadedManagedSettings = true
@@ -187,10 +152,6 @@
                 launchAtLogin: currentLaunchAtLoginPreference,
                 showInDock: currentShowInDockPreference
             )
-        }
-
-        func previewSound() {
-            InteractionSoundPlayer().playPreview(preset: .defaultPreset)
         }
 
         func restartOnboarding() {
