--- StetMac/Features/MacShell/HotKeySetting/MacHotkeySettingsView.swift
@@ -1,28 +1,0 @@
-#if os(macOS)
-    import SwiftUI
-
-    struct MacHotkeySettingsView: View {
-        @State private var message: String?
-
-        var body: some View {
-            Form {
-                Section {
-                    MacHotKeySettingsSectionView(hotkey: .dictation) { shortcut in
-                        message = shortcut.map { "Shortcut updated to \($0)." } ?? "Shortcut cleared."
-                    }
-                } header: {
-                    Text("Shortcut Engine")
-                }
-                if let message {
-                    Section {
-                        Text(message)
-                            .foregroundStyle(.secondary)
-                    }
-                }
-            }
-            .formStyle(.grouped)
-            .padding(.leading, MacUI.SettingsViewMetrics.formHorizontalPadding)
-            .padding(.bottom, MacUI.SettingsViewMetrics.formBottomPadding)
-        }
-    }
-#endif
