+++ StetMacTests/Features/MacShell/GeneralSetting/MacGeneralSettingsViewModelTests.swift
@@ -106,17 +106,5 @@
             #expect(viewModel.feedback?.message == "Dock icon enabled.")
             #expect(viewModel.feedback?.isError == false)
         }
-
-        @Test func dictationCompletionNotificationTogglePersists() {
-            let defaults = TestSupport.makeUserDefaults()
-            let viewModel = makeViewModel(defaults: defaults)
-
-            viewModel.load()
-            viewModel.dictationCompletionNotificationsEnabled = false
-
-            #expect(
-                defaults.object(forKey: MacPreferences.dictationCompletionNotificationsEnabled) as? Bool
-                    == false)
-        }
     }
 #endif
